Assignment Complete - Added Moderator Role

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,7 @@
 class PostsController < ApplicationController
 
   before_action :require_sign_in, except: :show
-  before_action :authorize_user, except: [:show, :new, :create]
+  before_action :authorize_user, except: [:show, :new, :create, :delete]
 
   def show
     @topic = Topic.find(params[:topic_id])
@@ -45,6 +45,12 @@ class PostsController < ApplicationController
    end
 
    def destroy
+     post = Post.find(params[:id])
+     unless current_user.moderator? || current_user.admin?
+       flash[:error] = "You must be a moderator or admin to do that."
+       redirect_to [post.topic, post]
+     end
+
      @post = Post.find(params[:id])
 
      if @post.destroy
@@ -64,7 +70,7 @@ class PostsController < ApplicationController
 
    def authorize_user
      post = Post.find(params[:id])
-     
+
      unless current_user == post.user || current_user.admin?
        flash[:error] = "You must be an admin to do that."
        redirect_to [post.topic, post]
